update cli, update catalog based on input
