+++ pkg/catalog/catalog.go
@@ -47,7 +47,7 @@ type Feature struct {
 	Description string  `json:"description" yaml:"description,omitempty" jsonschema:"description=Optional description of the feature,example=Provides advanced analytics and reporting capabilities"`
 	Billing     Billing `json:"billing" yaml:"billing" jsonschema:"description=Billing information for the feature"`
 	Audience    string  `json:"audience" yaml:"audience" jsonschema:"enum=public,enum=private,enum=beta,description=Intended audience for the feature,example=public"`
-	Usage       *Usage  `json:"usage,omitempty" yaml:"usage" jsonschema:"description=Usage limits granted by the feature"`
+	Usage       *Usage  `json:"usage,omitempty" yaml:"usage,omitempty" jsonschema:"description=Usage limits granted by the feature"`
 }
 
 // Usage defines usage limits granted by a feature.
